[511] Inline SideBar. Refactoring.
